pyproject.toml
[tool.poetry]
name = "nova-trader"
version = "0.3.0"
description = "Lean trading research, backtesting, and execution system"
authors = []
readme = "README.md"
packages = [
    { include = "src", from = "." },
]

[tool.poetry.dependencies]
python = "^3.11"

# Data & computation
pandas = "^2.1.0"
numpy = "^1.24.0"
requests = "^2.31.0"

# Models & validation
pydantic = "^2.4.2"

# Date handling
python-dateutil = "^2.8.0"

# Execution & data
alpaca-py = "^0.35.0"
yfinance = "^1.2.0"
# praw = "^7.7.1"  # Future: Reddit sentiment

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
force_alphabetical_sort_within_sections = true

[tool.poetry.scripts]
nova = "src.cli.app:main"
trade = "src.cli.app:main"
backtester = "src.cli.app:main"
